v2.1.0
Added Index65 Mode. This will print the Index65 hash to the terminal when encrypting and decrypting a file.
You may enable or disable this function by opening `aegis.bas` in a text editor and setting `index65Mode% = 1`
It is set to 0 by default.
This setting may only be changed globally. There is no command attribute.